Move status widgets to editor plugin widgets
